Fixed a lot of bugs related to attack, reduced player default attack speed and reach distance and implemented new fall damage system.

+++ src/main/java/tgw/evolution/items/IOffhandAttackable.java
@@ -2,7 +2,7 @@ package tgw.evolution.items;
 
 public interface IOffhandAttackable {
 
-    float getAttackSpeed();
+    double getAttackSpeed();
 
     double getReach();
